Move hook to separate file
