+++ src/typescript/frontend/src/lib/store/latest-balance/index.ts
@@ -1,0 +1,2 @@
+export * from "./store";
+export * from "./use-latest-balance";
